seperate trades
